app/requirements.txt
pollination-streamlit>=0.5.0
pollination-streamlit-io>=0.36.5
streamlit-ace>=0.1.1